% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COILplus-package.R
\docType{package}
\name{COILplus-package}
\alias{COILplus}
\alias{COILplus-package}
\title{COILplus: Covariate-Informed Latent Interaction Models for Bipartite Networks}
\description{
Implements the COIL and COIL+ models for predicting unobserved interactions in bipartite ecological networks using species trait information, phylogenetic correlations, and informative occurrence priors. COIL+ extends the original BiasedNetwork implementation (Papadogeorgou et al., 2023) with blocked sampling of occurrence probabilities and major efficiency improvements. Includes support for parallel trait-matching analysis via permutation tests.
}
\author{
\strong{Maintainer}: Jennifer Kampe \email{jenniferkampe@gmail.com}

}
