% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredLatFac_function.R
\name{GetPredLatFac}
\alias{GetPredLatFac}
\title{Samples from the distribution of the new latent factor.}
\usage{
GetPredLatFac(mcmc_latfac, mcmc_rho, C)
}
\arguments{
\item{mcmc_latfac}{Posterior samples from the distribution of latent factors
for species in the observed data.}

\item{mcmc_rho}{Posterior distribution of the weight in the latent factor
correlation matrix.}

\item{C}{Correlation matrix on species including the observed data species
and the one we wish to predict for.}
}
\value{
Matrix of with number of rows equal to the number of posterior
samples and number of columns equal to the number of latent factors.
}
\description{
Samples from the distribution of the new latent factor.
}
\keyword{internal}
