% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredWeights_function.R
\name{GetPredWeights}
\alias{GetPredWeights}
\title{Importance sampling weights for a set of units}
\usage{
GetPredWeights(pred_covs, pred_latfac, mcmc_coefs, mcmc_vars)
}
\arguments{
\item{pred_covs}{The covariates of the species we predict on. Can include
NAs for missing values.}

\item{pred_latfac}{The generated latent factors for this species.}

\item{mcmc_coefs}{Posterior samples for the coefficients in the covariate
models.}

\item{mcmc_vars}{Posterior samples for the residual variances in the models
for the continuous covariates.}
}
\description{
Corresponds to the likelihood of the species covariates given the generated
latent factors and current parameters.
}
\keyword{internal}
