% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmegaFromV_function.R
\name{OmegaFromV}
\alias{OmegaFromV}
\title{Getting omega from v in the increasing shrinkage prior.}
\usage{
OmegaFromV(v_val)
}
\arguments{
\item{v_val}{Vector of length equal to the value of H we use including
values between 0 and 1.}
}
\description{
Getting omega from v in the increasing shrinkage prior.
}
\keyword{internal}
