% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictInteractions_function.R
\name{PredictInteractions}
\alias{PredictInteractions}
\title{Predicting the probability of species interaction}
\usage{
PredictInteractions(pred_X, pred_W, mcmc, corr_mat)
}
\arguments{
\item{pred_X}{Matrix of covariates for the first set of species. Number of
rows are the number of measured covariates, and number of columns is the
number of species. The matrix can include missing data.}

\item{pred_W}{Matrix of covariates for the second set of species. Number of
rows are the number of measured covariates, and number of columns is the
number of species. The matrix can include missing data.}

\item{mcmc}{List of posterior samples as returned from the MCMC function.}

\item{corr_mat}{List of two elements. Each element corresponds to each set
of species, and it includes the species phylogenetic correlation matrix
extended to include the species on which we predict. For each species, we
only use the sub-matrix including the observed data and the target species.
If the species is included in the original data, it still has to be included
here, but its entries can be arbitrary, since they are not used.}
}
\value{
List including four arrays of dimensions corresponding to the
number of the two sets of species and posterior samples. The first array
includes draws from the probability of an interaction between the two
species based on the model, these probabilities are not in any way adjusted
by importance sampling weights. The second array includes draws from the
unweighted distribution without importance sampling reweighting. The third
array includes the importance sampling weights. The fourth array includes
samples from the posterior distribution of interactions after re-weighting
according to the importance sampling weights.
}
\description{
Using samples from the posterior distribution of model parameters to predict
the probability of interaction for in-sample co-occuring, in-sample
neglected, half-in-sample, and out-of-sample pairs of species.
}
\keyword{internal}
