% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitMatching_function.R
\name{TraitMatching}
\alias{TraitMatching}
\title{Trait matching for species interactions}
\usage{
TraitMatching(
  B = 500,
  mod_pL1s,
  Xs,
  Ws,
  obs_X,
  obs_W,
  obs_only = FALSE,
  ncores = 1
)
}
\arguments{
\item{B}{Integer. Number of permutation resamples. Default is 500.}

\item{mod_pL1s}{Array of posterior samples of fitted interaction
probabilities, as returned by \code{MCMC}. Dimensions: [Nsims, nB, nP].}

\item{Xs}{List of posterior imputed covariate values for the first set of
species, as returned by \code{MCMC}. Each element has rows equal to the
number of posterior samples and columns equal to the number of missing
observations for that covariate.}

\item{Ws}{Same as \code{Xs} but for the second set of species.}

\item{obs_X}{Design matrix of covariates for the first set of species
(species x covariates).}

\item{obs_W}{Design matrix of covariates for the second set of species
(species x covariates).}

\item{obs_only}{Logical. If \code{TRUE}, only species with fully observed
covariates are used. Defaults to \code{FALSE}.}

\item{ncores}{Integer. Number of cores for parallel processing over
posterior samples. Default is 1 (sequential). Values greater than 1
use a \code{doParallel} cluster.}
}
\value{
A list with elements:
\describe{
  \item{rsq_resampling_X}{Permutation R-squared for covariates of species
    type 1. Matrix [B x covariates].}
  \item{rsq_resampling_W}{Permutation R-squared for covariates of species
    type 2. Matrix [B x covariates].}
  \item{rsq_obs_X}{Observed mean R-squared for each covariate of species
    type 1.}
  \item{rsq_obs_W}{Observed mean R-squared for each covariate of species
    type 2.}
  \item{corr_obs_X}{Posterior mean correlation between each species type 2
    and each covariate of species type 1. Matrix [nP x covariates].}
  \item{corr_obs_W}{Posterior mean correlation between each species type 1
    and each covariate of species type 2. Matrix [nB x covariates].}
}
}
\description{
Fits univariate regressions of the logit of posterior interaction
probability samples on covariates and uses permutation to obtain the
null distribution of no association. Returns both R-squared (unsigned
importance) and mean posterior correlation (signed direction) for each
covariate.
}
