% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdLatFac_function_fast.R
\name{UpdLatFac_fast}
\alias{UpdLatFac_fast}
\title{Updating the model's latent factors
This version improves efficiency and accuracy by avoiding explicit inversions}
\usage{
UpdLatFac_fast(
  latfac,
  latfac_others,
  probobs,
  coefs_probobs,
  var_probobs,
  obs_covs,
  omega_obs_covs,
  num_covs,
  coefs_covs,
  var_covs,
  curr_inter,
  coefs_inter,
  omega_inter,
  prior_S_inv,
  cut_feed = FALSE
)
}
\arguments{
\item{latfac}{Current values of the latent factors to be updated. Matrix
with number of rows equal to the number of species and number of columns
equal to the number of factors.}

\item{latfac_others}{Current values of the latent factors fot the other set
of species, not to be updated. Matrix. Rows correspond to species and
columns correspond to latent factors.}

\item{probobs}{Vector including the values for the probability of observing
a given species interaction within a study. Should be of length equal to the
number of species among this set.}

\item{coefs_probobs}{Coefficients of the latent factors for the normal model
on the logit of probability of observing a given species. If these
coefficients are all NA, then this implies that no bias correction for
geographical and taxonomical bias is performed, so this part of the model
does not inform the updates of the latent factors. Should be of length equal
to the number of latent factors used plus 1 for the intercept.}

\item{var_probobs}{Numeric. Variance in the model for the probability of
detecting an interaction. Used only if performing bias correction.}

\item{obs_covs}{Matrix of observed traits. Rows correspond to species and
columns correspond to covariates. Continuous covariates should be included
first.}

\item{omega_obs_covs}{The Polya-Gamma omegas from the models of binary
covariates. Matrix. Rows correspond to units, columns to binary traits.}

\item{num_covs}{Vector of length 2 including the number of continuous and
binary traits.}

\item{coefs_covs}{Matrix of latent factor coefficients in the traits models.
Rows correspond to traits and columns to factors.}

\item{var_covs}{Vector of length equal to the number of continuous traits
including the residual variance of the corresponding model.}

\item{curr_inter}{Matrix with rows corresponding to the set of species whose
latent factors we are updating and columns corresponding to the other set of
species. Entries are 0 or 1.}

\item{coefs_inter}{Vector including the coefficients in the network model.}

\item{omega_inter}{The omegas generated from the Polya-Gamma for the model
of the true interactions L. The rows should correspond to the units whose
latent factors are updated using this function.}

\item{prior_S_inv}{The inverse of the current correlation matrix for the
latent factors that are being updated.}

\item{cut_feed}{Logical. Whether we should cut the feedback from the
interaction and detectability submodels into the latent factors. Defaults to
FALSE, in which case the full posterior is considered.}
}
\description{
MCMC sampling of latent factors for either set of species.
}
\keyword{internal}
