% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdProbObs_function.R
\name{UpdProbObs}
\alias{UpdProbObs}
\title{Updating the species probability of being observed in a study}
\usage{
UpdProbObs(
  probobs_curr,
  probobs_others,
  curr_inter,
  obs_inter,
  mh_n,
  obs_in_poss,
  unobs_in_poss,
  latfac,
  coefs_probobs,
  var_probobs
)
}
\arguments{
\item{probobs_curr}{Vector of current values for the probability of
observation for the set of species we are updating.}

\item{probobs_others}{Vector of current values for the probability of
observation for the other set of species, the ones that are not being
updated.}

\item{curr_inter}{Matrix with 0,1 entries for the current posterior samples
of possible interactions. The rows correspond to the species we are updating
and the columns correspond to the other set of species.}

\item{obs_inter}{Matrix with 0,1 entries for the detected interactions. The
rows correspond to the species we are updating and the columns correspond to
the other set of species.}

\item{mh_n}{Numeric. Parameter n in the Beta proposal distribution.}

\item{obs_in_poss}{Matrix with rows corresponding to the set of species we
are updating and rows corresponding to the other set. The matrix includes
a count for the number of studies that could have recorded the interaction
and did record it.}

\item{unobs_in_poss}{Matrix with rows corresponding to the set of species we
are updating and rows corresponding to the other set. The matrix includes
a count for the number of studies that could have recorded the interaction
but they did not record it.}

\item{latfac}{Current values of the latent factors for the set of species
whose probability of detection is to be updated. Matrix with rows
corresponding to the species and columns corresponding to the factors.}

\item{coefs_probobs}{Coefficients of the latent factors for the normal model
on the logit of probability of observing a given species. If these
coefficients are all NA, then this implies that no bias correction for
geographical and taxonomical bias is performed, so this part of the model
does not inform the updates of the latent factors. Should be of length equal
to the number of latent factors used plus 1 for the intercept.}

\item{var_probobs}{Numeric. Variance in the model for the probability of
detecting an interaction. Used only if performing bias correction.}
}
\description{
A species probabiliy of being observed, pi, is updated based on the observed
interactions, latent traits, and current true interactions.
}
\keyword{internal}
